Add number of time points as a parameter to find_bst_szr
- Previously, find_bst_szr() function is downsampling slp to 150 time
points. Updating the function so that we can pass the no. of time points
to which slp is downsampled to can be passed as a parameter

- Sensor power was computed directly from slp but it is causing seizures
with constant but high variance in baseline between sensors to be
selected over seizures with epileptor like transients in their log power
profile. Mean subtracting the slp can probably reduce this effect, so
mean subtracting slp before computing sensor power

- Also changed the function name from find_bst_szr() to
find_bst_szr_slp()