Update local coupling
The truncation of modes is producing negative values in local
coupling, although the convolution in time domain would be always
positive. These negative values are inhibiting the seizure
propagation, unless the truncation is high (lmax > 100). In order
to address this, the negative values are replaced with zeros.

Also, adding local coupling in dz/dt is producing seizures that
last very long. Moved the local coupling to dx/dt to address this.